Refactor user and project routes; implement user and project information retrieval, creation, and deletion logic
