+++ routes/user.routes.ts
@@ -1,3 +1,4 @@
+import { Router } from "express";
 import {
   signIN,
   registerUser,
